requirements.txt
#First create a virtual environment
#Activate it 
#pip install qdrant-client pypdf langchain langchain-community pyyaml requests
#ollama run llama3.2:3b  -> To install your LLm on your system
#ollama pull nomic-embed-text -> To install your embedding model
#ollama serve -. To just ollama
#Docker must be installed
#docker run -d -p 6335:6333 qdrant/qdrant -> To run qdrant
#http://localhost:6335/dashboard-> To access qdrant dashboard

qdrant-client
requests
streamlit
pyyaml
pypdf
langchain
langchain-community